refactor(cli): extract `drt run` to commands/run.py (#573 Phase 2b PR (a))
Moves the run command + its supporting infrastructure out of main.py:

- LogFormat Enum (#577)
- _STANDARD_LOG_FIELDS / _JsonFormatter / _configure_json_logging
- _RunContext dataclass
- _exit_code_for_signal (POSIX 128+signum)
- _run_one (per-sync execution; observer composition; telemetry)
- _print_watermark_summary (post-run notes)
- run (@app.command itself; signal handling; parallel/sequential dispatch)

Total ~485 LOC. cli/main.py shrinks 1086 → 595 (-45%).

Back-compat: main.py re-exports each moved underscore-prefixed name +
LogFormat so existing `from drt.cli.main import _RunContext` etc. keep
working. Same pattern as #572's `_get_*` shims.

Test mock paths updated for monkeypatched names that no longer live on
drt.cli.main:

- print_row_errors: was failing outright (AttributeError) — main.py no
  longer imports it after the move.
- _get_destination / _get_watermark_storage: mocks were silently no-ops
  since #572 because _run_one calls the public _helpers.get_destination
  directly; updating the path so the mock is actually effective.

Closes part of #573 (PR (b) covers validate / status / test).
